+++ src/tests/test_interface.py
@@ -572,3 +572,31 @@ def test_readme_compound():
   assert IFoo.provided_by(Foo())
   assert list(IFoo.implementations()) == [Foo]
   assert Foo(42).x == 42
+
+
+def test_implements_metaclass_conflict():
+  from nr.types.struct import Field, Struct
+  class MyInterface(Interface):
+    def foo(self):
+      pass
+
+  with pytest.raises(TypeError) as excinfo:
+    @implements(MyInterface, resolve_metaclass_conflict=False)
+    class MyImpl(Struct):
+      pass
+  assert 'metaclass conflict' in str(excinfo.value).lower()
+
+  with pytest.raises(ImplementationError) as excinfo:
+    @implements(MyInterface, resolve_metaclass_conflict=True)
+    class MyImpl(Struct):
+      pass
+  assert 'does not meet requirements of interface MyInterface' in str(excinfo.value)
+
+  @implements(MyInterface)
+  class MyImpl(Struct):
+    a = Field(int)
+    def foo(self):
+      return self.a * 2
+
+  assert MyImpl.__fields__
+  assert MyImpl(42).foo() == 84